feat(brand): marque unique « bee » + fix jours d'echeance + decision tablette (Phase 0)
- Marque : unification BimoTech / Bimotheque / Renlio -> « bee » partout
  (app interne, auth, pages legales, marketing, portail, PDF, emails onboarding,
  2FA, abonnement, erreurs, .env.testing). Identifiants de classe, cles config
  services.bimotech.*, emails @bimotech.sn et domaine bimotechsn.com inchanges.
- fix(ui) : jours d'echeance arrondis (int) -> contrats, etats-trimestriels,
  dashboard locataire (fini les « 19.133...j »).
- docs(claude) : 2 paliers responsive (mobile < 1024 / desktop), pas de chrome
  tablette dedie.

Prod : definir APP_NAME=bee dans le .env serveur (local deja fait).

+++ resources/views/admin/contrats/index.blade.php
@@ -159,7 +159,7 @@
                         {{ $contrat->date_debut?->format('d/m/Y') }} →
                         {{ $contrat->date_fin?->format('d/m/Y') ?? 'Ouvert' }}
                         @if($contrat->date_fin && $contrat->statut === 'actif')
-                            @php $jr = now()->diffInDays($contrat->date_fin, false); @endphp
+                            @php $jr = (int) now()->diffInDays($contrat->date_fin, false); @endphp
                             @if($jr <= 30 && $jr >= 0)
                                 <span class="text-bimo-gold font-semibold ml-1">⚠ {{ $jr }}j</span>
                             @endif
@@ -250,7 +250,7 @@
                         <td class="px-5 py-3.5 font-body text-xs text-bimo-text/60">
                             {{ $contrat->date_fin?->format('d/m/Y') ?? 'Ouvert' }}
                             @if($contrat->date_fin && $contrat->statut === 'actif')
-                                @php $jr = now()->diffInDays($contrat->date_fin, false); @endphp
+                                @php $jr = (int) now()->diffInDays($contrat->date_fin, false); @endphp
                                 @if($jr <= 30 && $jr >= 0)
                                     <div class="font-body text-[10px] text-bimo-gold font-semibold mt-0.5">⚠ {{ $jr }}j restants</div>
                                 @endif
